Change index of cfg nodes back to integer, entry nodes have index 0, exit nodes idx according to nodes size

+++ com.jdfc.core/src/main/java/data/MethodData.java
@@ -171,7 +171,7 @@ public class MethodData {
      * Calculates all possible Def-Use-Pairs.
      */
     public void calculateDefUsePairs() {
-        for (Map.Entry<Double, CFGNode> entry : this.cfg.getNodes().entrySet()) {
+        for (Map.Entry<Integer, CFGNode> entry : this.cfg.getNodes().entrySet()) {
             CFGNode node = entry.getValue();
 
             for (ProgramVariable def : node.getReach()) {
